+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-# 2025-08-03 (Version 0.3.0)
-Test package with external data, add more alias keywords for district in [parser_legacy.json](vietnamadminunits/data/parser_legacy.json)
+# 2025-08-03 (Version 0.4.0)
+- [parser_legacy.json](vietnamadminunits/data/parser_legacy.json): Add more alias keywords for district and ward level.
+- Add 0 for numeric districts and numeric wards. Eg: quan01, phuong01.
+- Fix wrong arg in [converter_2025.py](vietnamadminunits/converter/converter_2025.py): `nearest_point = find_nearest_point(a_point=old_point, list_of_b_points=new_ward_points)`
 
 # 2025-08-02 (Version 0.2.0)
 Public [vietnamadminunits](vietnamadminunits)